Data.Trade.GTIN

Description

The module exposes a GTIN data type that contains the number of digits as well.

GTIN and its aliasses.

newtype GTIN (n :: Natural)

A datatype for Global Trade Item Numbers GTIN with arbitrary "width" (up to nineteen digits technically possible).

Constructors

GTIN Word64

gtin

Arguments

:: forall i n. ((<=) n 19, Integral i, KnownNat n)

=> i
An Integral value for which we want to construct a GTIN number.

-> Maybe (GTIN n)
A GTIN number wrapped in a Just if the given value is within bounds and the checksum matches; Nothing otherwise.

Constructing a 'GTIN" with bound and checksum checks.

type GTIN14 = GTIN 14

A type alias for a GTIN number with fourteen numbers, with as range 00 0000 0000 0000–99 9999 9999 9997.

type GTIN13 = GTIN 13

A type alias for a GTIN number with thirteen numbers, with as range 0 0000 0000 0000–9 9999 9999 9994.

type GTIN12 = GTIN 12

A type alias for a GTIN number with twelve numbers, with as range 0000 0000 0000–9999 9999 9993.

type GTIN8 = GTIN 8

A type alias for a GTIN number with eight numbers, with as range 0000 0000–9999 9995.

type EANUCC14 = GTIN14

A type alias for a GTIN number with fourteen numbers, with as range 00 0000 0000 0000–99 9999 9999 9997.

type SCC14 = GTIN14

A type alias for a GTIN number with fourteen numbers, with as range 00 0000 0000 0000–99 9999 9999 9997.

type EAN = GTIN13

A type alias for a GTIN number with thirteen numbers, with as range 0 0000 0000 0000–9 9999 9999 9994.

type EANUCC13 = GTIN13

A type alias for a GTIN number with thirteen numbers, with as range 0 0000 0000 0000–9 9999 9999 9994.

type ISBN = GTIN13

A type alias for a GTIN with thirtheen numbers which is also an ISBN number, with as range 0 0000 0000 0000–9 9999 9999 9994.

type ISBN13 = GTIN13

A type alias for a GTIN with thirtheen numbers which is also an ISBN number, with as range 0 0000 0000 0000–9 9999 9999 9994.

type EANUCC8 = GTIN8

A type alias for a GTIN number with eight numbers, with as range 0000 0000–9999 9995.

type GSIN = GTIN 17

A type alias for a GTIN number with seventeen numbers, with as range 0 0000 0000 0000 0000–9 9999 9999 9999 9992.

type SSCC = GTIN 18

A type alias for a GTIN number with eighteen numbers, with as range 00 0000 0000 0000 0000–99 9999 9999 9999 9995.

Check if two GTINs are equivalent, even if the "width" of the GTINs are equivalent.

equivGTIN

Arguments

:: GTIN m
The first GTIN to check.

-> GTIN n
The second GTIN to check.

-> Bool
True if the given GTIN values are equivalent; False otherwise.

Check if two GTIN numbers, possibly with a different "width" are equivalent.

upscaleGTIN

Arguments

:: (<=) m n

=> GTIN m
The original GTIN number to upscale.

-> GTIN n
A GTIN with the same number, but more (or the same) number of digits.

Convert one GTIN into a GTIN that has more digits. The new GTIN will have additional leading zeros.

Fix the checksum of a GTIN number

fixChecksum

Arguments

:: GTIN n
The given GTIN number where we fix the checksum from.

-> GTIN n
A GTIN object that is the variant of the given GTIN number, with a valid checksum.

Fix the checksum of a given GTIN object. If the checksum is valid, then it will return the same GTIN, this operation is thus idempotent.

checkChecksum

Arguments

:: GTIN n
The given GTIN number for which we check the checksum.

-> Bool
True if the given checksum matches; False otherwise.

Check if the given checksum matches.

Convert the GTINs to a readable format.

gtinToString

Arguments

:: KnownNat n

=> GTIN n
The given GTIN number to convert to a readable String.

-> String
A String that contains the GTIN number, in chucks of four digits.

Convert the given GTIN number to convert to a String that groups numbers into groups of four.

ISBN-10 to ISBN-13

fromISBN10'

Arguments

:: Integral i

=> i
An Integral number that contains an ISBN-10.

-> ISBN13
The equivalent ISBN-13 number, which is a GTIN number with the corresponding checksum algorithm.

Convert a given integral number that contains an ISBN-10 number into the ISBN13 equivalent. For example 8175257660 is converted to 9 7881 7525 7665. This will add a 978 prefix, and recalculate the checksum.

Parsing GTINs

gtinParser

Arguments

:: forall s u m n. ((<=) 2 n, (<=) n 19, KnownNat n, Stream s m Char)

=> ParsecT s u m (GTIN n)
A parser parsing a GTIN with an arbitrary number of digits that forces the stream to end, and does checks the checksum.

A parser for a gtin number with an arbitrary number of digits between two and nineteen. the parser forces the stream to end after the gtin, and validates if the gtin is indeed valid. The parser parses the number of digits with an arbitrary number of spaces between any two digits.

gtinParser_

Arguments

:: forall s u m n. ((<=) 2 n, (<=) n 19, KnownNat n, Stream s m Char)

=> ParsecT s u m (GTIN n)
A parser parsing a GTIN with an arbitrary number of digits that does not force the stream to end, but checks the checksum.

A parser for a gtin number with an arbitrary number of digits between two and nineteen. the parser does not end after the gtin (so no eof is required). The GTIN is validated, so if the checksum does not match, the parser fails. The parser parses the number of digits with an arbitrary number of spaces between any two digits.

gtinParser'

Arguments

:: forall s u m n. ((<=) 2 n, (<=) n 19, KnownNat n, Stream s m Char)

=> ParsecT s u m (GTIN n)
A parser parsing a GTIN with an arbitrary number of digits thats force the stream to end, but does not check the checksum.

A parser for a gtin number with an arbitrary number of digits between two and nineteen. the parser forces the stream to end after the gtin, but does not validate if the gtin is indeed valid. The parser parses the number of digits with an arbitrary number of spaces between any two digits.

gtinParser_'

Arguments

:: forall s u m n. ((<=) 2 n, (<=) n 19, KnownNat n, Stream s m Char)

=> ParsecT s u m (GTIN n)
A parser parsing a GTIN with an arbitrary number of digits that does not force the stream to end, and does not check the checksum.

A parser for a gtin number with an arbitrary number of digits between two and nineteen. the parser does not end after the gtin (so no eof is required), and furthermore does not validate if the gtin is indeed valid. The parser parses the number of digits with an arbitrary number of spaces between any two digits.

parseGTIN

Arguments

:: forall n s. ((<=) 2 n, (<=) n 19, KnownNat n, Stream s Identity Char)

=> s
The stream to parse.

-> Either ParseError (GTIN n)
The result of the parser: Either a ParseError or the parsed GTIN.

Run the gtinParser_' parser and thus parses a GTIN with an arbitrary number of digits. The parser requires the stream to end after the GTIN, and validates the checksum.

parseGTIN_

Arguments

:: forall n s. ((<=) 2 n, (<=) n 19, KnownNat n, Stream s Identity Char)

=> s
The stream to parse.

-> Either ParseError (GTIN n)
The result of the parser: Either a ParseError or the parsed GTIN.

Run the gtinParser_' parser and thus parses a GTIN with an arbitrary number of digits. The parser does not require the stream to end after the GTIN, but validates the checksum.

parseGTIN'

Arguments

:: forall n s. ((<=) 2 n, (<=) n 19, KnownNat n, Stream s Identity Char)

=> s
The stream to parse.

-> Either ParseError (GTIN n)
The result of the parser: Either a ParseError or the parsed GTIN.

Run the gtinParser_' parser and thus parses a GTIN with an arbitrary number of digits. The parser requires the stream to end after the GTIN, but does not validate the checksum.

parseGTIN_'

Arguments

:: forall n s. ((<=) 2 n, (<=) n 19, KnownNat n, Stream s Identity Char)

=> s
The stream to parse.

-> Either ParseError (GTIN n)
The result of the parser: Either a ParseError or the parsed GTIN.

Run the gtinParser_' parser and thus parses a GTIN with an arbitrary number of digits. The parser does not require the stream to end after the GTIN, and does not validate the checksum.

QuasiQuoters

gtinQ

Arguments

:: forall n. ((<=) 2 n, (<=) n 19, KnownNat n)

=> Proxy (GTIN n)
The Proxy object that is ignored, but used to determine the number of digits of the GTIN number.

-> QuasiQuoter
The corresponding QuasiQuoter that resolves to a GTIN expression or pattern.

A function that constructs a GTIN expression or pattern based on a given string. The Proxy parameter is used to specify the number of digits of the GTIN number.

gtin14Q

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an GTIN14 number.

gtin13Q

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an GTIN13 number.

gtin12Q

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an GTIN12 number.

gtin8Q

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an GTIN8 number.

eanucc8Q

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an EANUCC8 number.

eanucc14Q

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an EANUCC14 number.

scc14Q

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an SCC14 number.

eanQ

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an EAN number.

eanucc13Q

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an EANUCC13 number.

gsinQ

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for a GSIN number.

ssccQ

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an SSCC number.

isbnQ

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an ISBN number, only the thirteen digit number is supported.

isbn13Q

Arguments

:: QuasiQuoter
The corresponding QuasiQuoter.

A QuasiQuoter for an ISBN13 number.